requirements.txt
requests==2.31.0
feedgen==0.9.0
PyYAML==6.0.1
pyyaml-include==1.3
python-dateutil==2.9.0.post0 